infra/datalake/requirements.txt
apache-airflow-providers-apache-spark==5.4.0
apache-airflow-providers-docker==4.5.1
pyspark==4.0.1